requirements.txt
ipdb
pandas
requests
pyyaml
tqdm
